Add support for building a squashed initramfs
With all files stored in ramfs, and most of them are not compressed,
the initramfs will take up a lot of memory. This becomes much worse
with larger page size, as page cache will allocate a minimum of one
page to cache the content of one file however small it is.

One approach to reducing the memory usage and shrink initramfs size
is to put most files into a read-only squash image and keep a minimum
set of executable and libraries outside as the loader for the squash
image. Once the squash image is mounted, the real init inside will be
executed and then everything behaves as usual.

This patch will introduce a 99squash module which will never be included
by default. User can force add it, and if it is included, dracut
will perform some extra steps before creating the final image.

For now, /etc and /usr will be moved into the squashfs image. /init
will be renamed to /init.orig and replaced by /init.squash.
Files and folders need to be accessible before mounting the image will
be kept at their original place, not moving into the squashfs. A symlink
will be created inside the squashfs instead, pointing to the file
location at boot time, so no redundant binary or libraries are added.

/init.squash is now the init loader of the initramfs. It will do some
mounting and bind-mounting job, and call the original /init.orig,
then everything work as usual.

This is very helpful when mem resource is very limited, like Kdump.
According to my tests, with initramfs built with mkdumprd (which is
already a smaller image than usual), this can save about 40MB of
memory with 64K page size, or about 20MB with 4K page size.

Won't change any behavior if 99squash module is not enabled.