dracut.spec: reintroduce 51-dracut-rescue-postinst.sh script
Commit 0bb9a683d4d ("spec: drop support for legacy distributions") dropped
the 51-dracut-rescue-postinst.sh script, but this is used by Fedora during
installation to create the kernel and initramfs images and on a new kernel
install if these images don't exist, when using grubby (which seems to be
the case on a Fedora system upgrade).

So this commit partially reverts 0bb9a683d4d, to bring the script again.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>