dracut-install: Allow installing files from a sysroot
Introduced option -r SYSROOTDIR to use as a prefix for
source files.

Instead of relying on realpath(2), a new get_real_file() function
was introduced. This is to fix cases where symlinks inside the
sysroot are invalid on the build host. They would become valid
when running on the target system.

Most of the file checking operations work on source files instead
of the ones already copied to destination.

Coupled with a cross-compiler ldd script, library dependency
resolution works both in -R/--resolvelazy and -l/--ldd modes.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>