01bootetc: introduce module for updating dracut etc
This provides support for updating '/etc/' in the initramfs using
'bootetc=<DEVICE>'. If the device has the path '/initrd-etc', then the
tree is copied into the initramfs.

The purpose of bootetc is to aide immutable operating systems like
Fedora and Red Hat CoreOS that ship static initrds with a path for
user-provided configuration. Example use cases include udev rules, iSCSI
and multipath configuration, fips and Network Manager configurations.