dracut: 99shutdown fails if systemctl reboot does not go through
Problem:
In 99shutdown, it is possible for $ACTION -f -d -n to fail for reboot/poweroff,
in a situation where the $ACTION depends on systemctl D-Bus to be available.

At this point in the shutdown process, all processes/services are down, all FSs
are umount'd and all services that dictate .need_shutdown are downed by the
_check_shutdown loop. All that remains is to reboot|poweroff|halt the system
and there is no remaining sync, umount, or other shutdown work to be done.

Fix:
Change $ACTION -f -d -n to use systemctl $ACTION -ff to --force --force the
$ACTION. This will bypass any requirement on D-Bus and kick the system in gear.

Closes #869. See #869 for details.