40network: Keep nameservers from all interfaces
Make sure that all nameservers configured by each interface are kept in
/etc/resolv.conf. The current code will only keep the nameservers
configured for the latest interface. To preserve the existing order for
compatibility, we make sure to keep the last added nameservers first.

This matters especially in the case where the latest interface is
configured via DHCP and if the DHCP server does not send a nameserver.
The current code will remove all previously configured nameservers from
/etc/resolv.conf leaving the system unable to resolve hostnames.