Drop 51-dracut-rescue-postinst.sh entirely
It should not be installed in RHEL-8. See #975 and
https://bugzilla.redhat.com/show_bug.cgi?id=1677723 for further
detail; basically this should only ever be installed on RHEL < 8
and Fedora < 31, but it got pinged around a lot and the RHEL-8
branch happened to be forked off at a point where it gets
installed unconditionally. Later changes to make it only install
on RHEL < 8 were committed to master but never to the RHEL-8
branch, so never really did anything. Since this branch is for
RHEL-8 specifically, there's no sense fiddling with conditionals,
we can just ditch it entirely.

Signed-off-by: Adam Williamson <awilliam@redhat.com>