feat(dracut): add --preferred-init option
Introduce 'preferred-init' option to replace the default init program
with a customized one, while moving the default /init to /init.dracut

    Usage: dracut --preferred-init /path-to/customized-init.sh

The init program may need customizations when multiple boot scenarios
are possible on a system. Along with this option, --include could be
used to include squash'ed image(s) built for the other possible boot
scenario(s). And the customized init program can be leveraged to exec
into the default init sequence (exec /init.dracut) in case of regular
boot scenario and to activate the squash'ed image in case of the other
possible boot scenario(s).

Signed-off-by: Hari Bathini <hbathini@linux.ibm.com>