fix(dracut.sh): require root to run hostonly mode
The base module install function (modules.d/99base/module-setup.sh) cannot
read /etc/shadow in hostonly mode if the user is not root, but dracut allows
to create the initramfs anyway. This causes that it is not possible to log in to
the emergency shell with the newly created initramfs.