fix(udev-rules): make sure the hwdb with ifname renaming rules is around
The stable network interface naming relies on hwdb. In one particular
case actually: the Dell "idrac" interface.

The systemd-udevd module actually already includes the hwdb for hostonly
images, but we need it to work for non-hostonly images as well as those
that don't utilize the systemd-udevd module as well.

Let's make udev-rules also pull in the binary hwdb for hostonly builds.
On Fedora 35 this adds 2M compressed, 11M uncompressed. That is quite
some extra fat, but it's consistent with what systemd-udevd does.

For non-hostonly ones, pull in only what we know we need in binary form.
The extra space consumed is negligible (300B uncompressed).