refactor(base): use blockdev@<devname>.target instead of <devname>.device
The blockdev@<devname>.target is a unit made for ordering block devices
for systemd services. It seems to be the canonical way of doing this as
given in the man page systemd.special. This PR aims to get dracut to
use the same target uint throughout in base, as well as in the
dracut-systemd module.

Signed-off-by: Savyasachee Jha <hi@savyasacheejha.com>