fix(dracut-systemd): systemd-gpt-auto-generator should not run when root is unset
systemd-gpt-auto generator runs when `root=<device>` is not set on the
kernel cmdline. It doesn't work even if `root=<device>` is set by dracut
in the stored command line. For systemd-gpt-auto-generator to not run,
one either needs to explicitly set a `root=` parameter or set
`rd.systemd.gpt_auto=0` on the kernel cmdline.

Since we do not have control over the command line passed to the kernel,
the easiest way to bypass this is to delete the generator when the root
parameter is set.

Fixes #1709

Signed-off-by: Savyasachee Jha <hi@savyasacheejha.com>