fix: Make NetworkManager check interfaces repeatedly until up
An interface may come up late, after the carrier timeout has elapsed and
after nm-wait-online-initrd.service has completed; for instance, DHCP
may be slow. Previously, the nm-run settled hook checked each interface
only once for whether it is up, and if an interface came up after the
nm-run settled hook was run, then the online hook would not be run for
that interface.

Now, the nm-run settled hook will run repeatedly to check for newly
activated NM-managed connections, then run the online hooks on the
devices of that connection on which the online hooks have not yet been
run. The nm-run settled hook is done once all connections generated by
nm-initrd-generator are activated and the online hook is run for those
connections' devices.