feat(multipath): warn if included with no multipath devices and no user conf
The multipath module is always included in no-hostonly mode, but including it in
the initramfs might activate multipath on the root device, which could be
undesirable. This behavior can be manually fixed by explicitly excluding the
root device from multipath in /etc/multipath.conf.

Therefore, issue a warning in no-hostonly mode, if there is no multipath device
and the user has not configured multipath.

Signed-off-by: Martin Wilck <mwilck@suse.de>
Signed-off-by: Antonio Alvarez Feijoo <antonio.feijoo@suse.com>