feat(lvm): install lvm commands symlinks
It will install for convenience, lvm's core commands tools.

Lvm package installs all of them as symlink to the binary,
and they are universally used without `lvm` command. If someone
is reaching for these tools in the initrd it's probably because
something is wrong, and so the admin may be stressed or flustered
and may not connect that "lvm" as a command does everything they
need. So recreating a familiar environment during the emergency
shell will really help them.

Signed-off-by: Valentin Lefebvre <valentin.lefebvre@suse.com>