Guard kernel install hook
Only run our kernel-install if the initrd generator is set to dracut,
and if the layout asks for a conventional bls layout.

In particular do not run this hook if the initrd generator is
mkinitcpio, as used by Arch's standard mkinitpcio tool, and do not run
this hook if the layout is set to uki, i.e. if the user asks for a
unified kernel image to be generated.